refactor: refactor helpres file
