+++ src/albums/albums.module.ts
@@ -1,10 +1,10 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AlbumEntity } from '../helpers';
 import { ArtistsModule } from '../artists/artists.module';
 import { ArtistsService } from '../artists/artists.service';
 import { AlbumsController } from './albums.controller';
 import { AlbumsService } from './albums.service';
+import { AlbumEntity } from './entities/album.entity';
 
 @Module({
   imports: [
