+++ src/artists/artists.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ArtistEntity } from '../helpers';
 import { ArtistsController } from './artists.controller';
 import { ArtistsService } from './artists.service';
+import { ArtistEntity } from './entities/artist.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([ArtistEntity])],
