+++ src/favorites/favorites.module.ts
@@ -1,6 +1,5 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FavoritesEntity } from '../helpers';
 import { AlbumsModule } from '../albums/albums.module';
 import { ArtistsModule } from '../artists/artists.module';
 import { TracksModule } from '../tracks/tracks.module';
@@ -9,6 +8,7 @@ import { FavoritesService } from './favorites.service';
 import { TracksService } from 'src/tracks/tracks.service';
 import { AlbumsService } from 'src/albums/albums.service';
 import { ArtistsService } from 'src/artists/artists.service';
+import { FavoritesEntity } from './entities/favorites.entity';
 
 @Module({
   imports: [
