+++ src/tracks/tracks.module.ts
@@ -1,12 +1,12 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { TrackEntity } from '../helpers';
 import { AlbumsModule } from '../albums/albums.module';
 import { AlbumsService } from '../albums/albums.service';
 import { ArtistsModule } from '../artists/artists.module';
 import { ArtistsService } from '../artists/artists.service';
 import { TracksService } from './tracks.service';
 import { TracksController } from './tracks.controller';
+import { TrackEntity } from './entities/track.entity';
 
 @Module({
   imports: [
